requirements.txt
requests==2.32.3
2captcha-python==1.5.1
phonenumbers==9.0.0
playwright==1.50.0
python-dotenv==1.0.1